+++ docs/CHANGES.rst
@@ -12,6 +12,7 @@ Change Log
   * ``nemo_cmd.prepare.check_xios_exec()``
   * ``nemo_cmd.prepare.copy_run_set_files()``
   * ``nemo_cmd.prepare.make_executable_links()``
+  * ``nemo_cmd.prepare.make_grid_links()``
   * ``nemo_cmd.prepare.make_namelists()``
   * ``nemo_cmd.prepare.make_run_dir()``
   * ``nemo_cmd.prepare.remove_run_dir()``